pyproject.toml
[project]
name = "aiwrite"
version = "0.1.0"
description = "自动化论文写作流水线：从题目+大纲到LaTeX/Word论文"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AIWrite Team"}
]
keywords = ["llm", "paper-writing", "latex", "automation"]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "rich>=13.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
aiwrite = "aiwrite.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aiwrite"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
